Create Machine Learning Algorithm
This is the Machine Learning Algorithm code. You upload the data you collect from the Python Obstacle Course to this. Then, the algorithm will interpret the data and make a prediction. The best results occur with as much data as possible. If you would like to use a data set, I provided the combined data I got from the Python Obstacle Course in 'combined_data(so far).csv.'